refactor: split dom.js into domain-specific modules
Split the dom.js hotspot (imported by 49+ files) into three
focused sub-modules by responsibility:

- dom-element.js    : _el, createActionButton
- dom-rendering.js  : renderButtonBar, buildDomainButtonBar, renderList, buildChevronRow
- dom-visibility.js : _vis

dom.js is preserved as a barrel re-export so all existing consumers
continue working without import-path changes.

Closes #541
